Move resolveEns logic to helpers
